backend/engine/requirements.txt
pydantic
pyyaml
colorlog

langchain
langchain_community
langchain_chroma
langchain_huggingface
langchain_ollama


ollama

wikipedia

pymupdf